Tweaked on General Algorithms

+++ utils/RCBot2_meta/bot_ga.cpp
@@ -35,44 +35,43 @@
 #include "bot_ga.h"
 #include "bot_globals.h"
 
-
-const int CGA :: g_iDefaultMaxPopSize = 16;
-const float CGA :: g_fCrossOverRate = 0.7f;
-const float CGA :: g_fMutateRate = 0.1f;
-const float CGA :: g_fMaxPerturbation = 0.3f;
+const int CGA::g_iDefaultMaxPopSize = 16;
+const float CGA::g_fCrossOverRate = 0.7f;
+const float CGA::g_fMutateRate = 0.1f;
+const float CGA::g_fMaxPerturbation = 0.3f;
 
 ////////////////////
 // POPULATION
 ////////////////////
 
-IIndividual *CPopulation :: get ( int iIndex ) const
+IIndividual* CPopulation::get(unsigned int iIndex) const
 {
 	return m_theIndividuals[iIndex];
 }
 
-void CPopulation :: add ( IIndividual *individual )
-{	
+void CPopulation::add(IIndividual* individual)
+{
 	m_theIndividuals.emplace_back(individual);
 }
 
-void CPopulation :: freeMemory ()
+void CPopulation::freeMemory()
 {
-	for ( unsigned int i = 0; i < m_theIndividuals.size(); i ++ )
+	for (unsigned int i = 0; i < m_theIndividuals.size(); i++)
 		delete m_theIndividuals[i];
 
 	m_theIndividuals.clear();
 }
 
-void CPopulation :: clear ()
+void CPopulation::clear()
 {
 	m_theIndividuals.clear();
 }
 
-ga_nn_value CPopulation :: totalFitness () const
+ga_nn_value CPopulation::totalFitness() const
 {
 	float fTotalFitness = 0.0f;
 
-	for ( unsigned int i = 0; i < size(); i ++ )
+	for (unsigned int i = 0; i < size(); i++)
 	{
 		fTotalFitness += m_theIndividuals[i]->getFitness();
 	}
@@ -80,16 +79,16 @@ ga_nn_value CPopulation :: totalFitness () const
 	return fTotalFitness;
 }
 
-ga_nn_value CPopulation :: bestFitness () const
+ga_nn_value CPopulation::bestFitness() const
 {
 	BOOL gotBestFitness = FALSE;
 	float fBestFitness = 0.0f;
 
-	for ( unsigned int i = 0; i < size(); i ++ )
+	for (unsigned int i = 0; i < size(); i++)
 	{
 		const float fFitness = m_theIndividuals[i]->getFitness();
 
-		if ( !gotBestFitness || fFitness > fBestFitness )
+		if (!gotBestFitness || fFitness > fBestFitness)
 		{
 			fBestFitness = fFitness;
 			gotBestFitness = TRUE;
@@ -99,14 +98,14 @@ ga_nn_value CPopulation :: bestFitness () const
 	return fBestFitness;
 }
 
-ga_nn_value CPopulation :: averageFitness () const
+ga_nn_value CPopulation::averageFitness() const
 {
-	return totalFitness()/m_theIndividuals.size();
+	return totalFitness() / static_cast<float>(m_theIndividuals.size());
 }
 
-IIndividual *CPopulation :: pick ()
+IIndividual* CPopulation::pick()
 {
-	IIndividual *to_return = m_theIndividuals.back();
+	IIndividual* to_return = m_theIndividuals.back();
 
 	m_theIndividuals.pop_back();
 
@@ -117,45 +116,45 @@ IIndividual *CPopulation :: pick ()
 // GENETIC ALGORITHM
 ////////////////////
 
-CGA :: CGA (ISelection *selectFunction) : m_theSelectFunction (selectFunction)
+CGA::CGA(ISelection* selectFunction) : m_theSelectFunction(selectFunction)
 {
 	init();
 	m_thePopulation.clear();
 	m_theNewPopulation.clear();
 }
 
-void CGA :: addToPopulation ( IIndividual *individual )
+void CGA::addToPopulation(IIndividual* individual)
 {
 	m_thePopulation.add(individual);
 
-	if ( m_thePopulation.size() >= m_iMaxPopSize )
+	if (m_thePopulation.size() >= m_iMaxPopSize)
 	{
 		epoch();
 		m_thePopulation.freeMemory();
 	}
 }
 
-void CGA :: epoch ()
+void CGA::epoch()
 {
 	m_theNewPopulation.freeMemory();
 
-	while ( m_theNewPopulation.size() < m_iMaxPopSize )
+	while (m_theNewPopulation.size() < m_iMaxPopSize)
 	{
-		IIndividual *mum = m_theSelectFunction->select(&m_thePopulation);
-		IIndividual *dad = m_theSelectFunction->select(&m_thePopulation);
+		IIndividual* mum = m_theSelectFunction->select(&m_thePopulation);
+		IIndividual* dad = m_theSelectFunction->select(&m_thePopulation);
 
-		IIndividual *baby1 = mum->copy();
-		IIndividual *baby2 = dad->copy();
+		IIndividual* baby1 = mum->copy();
+		IIndividual* baby2 = dad->copy();
 
-		if ( randomFloat(0,1) < g_fCrossOverRate )	
+		if (randomFloat(0, 1) < g_fCrossOverRate)
 			baby1->crossOver(baby2);
 
 		baby1->mutate();
 		baby2->mutate();
-		
+
 		m_theNewPopulation.add(baby1);
 		m_theNewPopulation.add(baby2);
-	}	
+	}
 
 	m_iNumGenerations++;
 
@@ -173,26 +172,26 @@ void CGA :: epoch ()
 	m_fPrevAvgFitness = fCurAvgFitness;
 }
 
-void CGA :: freeLocalMemory ()
+void CGA::freeLocalMemory()
 {
 	m_thePopulation.freeMemory();
 	m_theNewPopulation.freeMemory();
 	m_iNumGenerations = 0;
 }
 
-void CGA :: freeGlobalMemory ()
+void CGA::freeGlobalMemory()
 {
 	freeLocalMemory();
 	delete m_theSelectFunction;
 	m_theSelectFunction = nullptr;
 }
 
-bool CGA :: canPick () const
+bool CGA::canPick() const
 {
 	return m_theNewPopulation.size() > 0;
 }
 
-IIndividual *CGA :: pick ()
+IIndividual* CGA::pick()
 {
 	return m_theNewPopulation.pick();
 }
@@ -201,18 +200,18 @@ IIndividual *CGA :: pick ()
 // SELECTION
 ///////////////
 
-IIndividual *CRouletteSelection :: select ( CPopulation *population )
+IIndividual* CRouletteSelection::select(CPopulation* population)
 {
-	const ga_nn_value fFitnessSlice = randomFloat(0,population->totalFitness());
+	const ga_nn_value fFitnessSlice = randomFloat(0, population->totalFitness());
 	ga_nn_value fFitnessSoFar = 0.0f;
 
-	for ( unsigned int i = 0; i < population->size(); i ++ )
+	for (unsigned int i = 0; i < population->size(); i++)
 	{
-		IIndividual *individual = population->get(i);
+		IIndividual* individual = population->get(i);
 
 		fFitnessSoFar += individual->getFitness();
 
-		if ( fFitnessSoFar >= fFitnessSlice )
+		if (fFitnessSoFar >= fFitnessSlice)
 			return individual;
 	}
